+++ src/analyze/src/store/modules/configuration.js
@@ -7,6 +7,7 @@ const state = {
   loaderInFocusConfiguration: {},
   extractorInFocusConfiguration: {},
   extractorInFocusEntities: {},
+  pipelines: [],
 };
 
 const getters = {
@@ -17,6 +18,9 @@ const getters = {
   },
   getExtractorNameWithoutPrefixedTapDash() {
     return extractor => extractor.replace('tap-', '');
+  },
+  getHasPipelines() {
+    return state.pipelines.length > 0;
   },
   getLoaderImageUrl(_, gettersRef) {
     return loader => (
@@ -39,6 +43,13 @@ const actions = {
 
   clearLoaderInFocusConfiguration({ commit }) {
     commit('setLoaderInFocusConfiguration', {});
+  },
+
+  getAllPipelineSchedules({ commit }) {
+    orchestrationsApi.getAllPipelineSchedules()
+      .then((response) => {
+        commit('setPipelines', response.data);
+      });
   },
 
   getExtractorInFocusEntities({ commit }, extractorName) {
@@ -81,6 +92,13 @@ const actions = {
     orchestrationsApi.savePluginConfiguration(configPayload);
     // TODO commit if values are properly saved, they are initially copied from
     // the loader's config and we'd have to update this
+  },
+
+  savePipelineSchedule({ commit }, pipelineSchedulePayload) {
+    orchestrationsApi.savePipelineSchedule(pipelineSchedulePayload)
+      .then((response) => {
+        commit('updatePipelines', response.data);
+      });
   },
 
   selectEntities() {
@@ -153,8 +171,16 @@ const mutations = {
     state.loaderInFocusConfiguration = configuration;
   },
 
+  setPipelines(_, pipelines) {
+    state.pipelines = pipelines;
+  },
+
   toggleSelected(_, selectable) {
     Vue.set(selectable, 'selected', !selectable.selected);
+  },
+
+  updatePipelines(_, pipeline) {
+    state.pipelines.push(pipeline);
   },
 };