+++ src/analyze/src/utils/utils.js
@@ -115,4 +115,30 @@ export default {
   underscoreToSpace(value) {
     return value.replace(/_/g, ' ');
   },
+
+  // Date Utils
+  getTodayYYYYMMDD() {
+    return this.formatDateYYYYMMDD(new Date());
+  },
+
+  getIsDateStringInFormatYYYYMMDD(dateString) {
+    const result = /[0-9]{4}-[0-9]{2}-[0-9]{2}/.test(dateString);
+    return result;
+  },
+
+  formatDateYYYYMMDD(date) {
+    const d = new Date(date);
+    let month = `${d.getMonth() + 1}`;
+    let day = `${d.getDate()}`;
+    const year = d.getFullYear();
+
+    if (month.length < 2) {
+      month = `0${month}`;
+    }
+    if (day.length < 2) {
+      day = `0${day}`;
+    }
+
+    return [year, month, day].join('-');
+  },
 };